Enable Docker AND find a usable image
The below images aren't compatible with Nextflow as far as I can tell
due to Nextflow forcing `/bin/bash` as the entrypoint. This does not
seem to be adjustable - I tried through setting shell in the config and
also the process to no avail.

* docker.io/busybox
* docker.io/alpine
* docker.io/bash

Decided to stick with the Ubuntu image on the public ECR.